requirements.txt
numpy==1.26.4
pandas==2.2.3
opencv-python==4.10.0.82
cvlib==0.2.7
matplotlib==3.8.4
Pillow==10.4.0
tensorflow==2.16.1
scikit-learn==1.2.2
